+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This version also includes some major internal modifications on the Geo module.
 * Linear RGB interpolation now also interpolates alpha channel
 * bug #202 fix an issue in Stack when all series do not contains values for each category
 * bug #204 BandedScale padding and alignment now use Percent
+* bug #216 fix an issue when a scale returned tick values outside of its range
 
 ## 0.8.0-RC11
 > Published 14 Jul 2020